+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Versioning].
 
 ### Features
 
+- Maintain capitalization of all words in a mapping phrase. ([#26])
 - Add option to specify a mapping from the CLI. ([#31])
 - Replace phrases with spaces if a match is found with a newline. ([#37])
 
@@ -53,6 +54,7 @@ Versioning].
 [#22]: https://github.com/ericcornelissen/wordrow/pull/22
 [#21]: https://github.com/ericcornelissen/wordrow/pull/21
 [#23]: https://github.com/ericcornelissen/wordrow/pull/23
+[#26]: https://github.com/ericcornelissen/wordrow/pull/26
 [#29]: https://github.com/ericcornelissen/wordrow/pull/29
 [#30]: https://github.com/ericcornelissen/wordrow/pull/30
 [#31]: https://github.com/ericcornelissen/wordrow/pull/31